+++ lib/src/constants/mp_constants.dart
@@ -406,6 +406,7 @@ const String mpScrapFreeTextTHID = 'FREE_TEXT';
 const String mpUnknownPLAType = 'unknown';
 const String mpPLATypeSubtypeSeparator = ':';
 const String mpSubtypeIDSeparator = '|';
+const String mpToUserOptionsSeparator = ' | ';
 const String mpPointSubtypeIDPrefix = 'point$mpSubtypeIDSeparator';
 const String mpLineSubtypeIDPrefix = 'line$mpSubtypeIDSeparator';
 const String mpAreaSubtypeIDPrefix = 'area$mpSubtypeIDSeparator';