+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * New features:
   * Changed "Create map connection line from point X-Section to Station" shortcut from Ctrl+X to Ctrl+Alt+X.
   * Including an ID on automatically added map connection lines.
+  * Translating option names on status bar.
 * Fixed bugs:
   * "Run Therion" button blinking red on every start.
 * Infrastructure maintenance: